use FloatType in proc worker throughout (no longer storing uint16 rf)

+++ ArpamGuiQt/src/DataProcWorker.hpp
@@ -18,7 +18,7 @@ namespace fs = std::filesystem;
  * From RF to Image
  */
 template <uspam::Floating FloatType> struct BScanData {
-  arma::Mat<uint16_t> rf;
+  arma::Mat<FloatType> rf;
   uspam::io::PAUSpair<FloatType> rfPair;
   uspam::io::PAUSpair<uint8_t> rfLog;
 };
